backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
python-docx==1.2.0
pdfplumber==0.11.9
pydantic==2.12.5
